Flytter DELETE funksjon til contect/aktivbruker
